Specular probe checkpoint: * Specular probes mostly implemented.  Currently only the global probe is used, though.  Still need to generate mip-maps, apply pre-filtering, and use local probes with parallax correction. * Offloaded light culling management onto the light list object instead of the renderer. * Clustered lighting now has an adjustable tile size.  Each tile is still a square, but the largest dimension of the viewport is sized ~16 tiles wide with the tile size being a multiple of 8.
--Misc changes--
* Added material swaps for scene objects.
* Added StringCache utility to simplify string management.
* Moved some global debug state values into the user config file so they can be changed without recompiling.

+++ src/RenderLab/render/RdrPostProcess.cpp
@@ -428,7 +428,7 @@ void RdrPostProcess::DoTonemap(RdrContext* pRdrContext, RdrDrawState& rDrawState
 	RdrRenderTargetView renderTarget = RdrResourceSystem::GetRenderTargetView(RdrResourceSystem::kPrimaryRenderTargetHandle);
 	RdrDepthStencilView depthView = { 0 };
 	pRdrContext->SetRenderTargets(1, &renderTarget, depthView);
-	pRdrContext->SetViewport(Rect(0.f, 0.f, pColorBuffer->texInfo.width, pColorBuffer->texInfo.height));
+	pRdrContext->SetViewport(Rect(0.f, 0.f, (float)pColorBuffer->texInfo.width, (float)pColorBuffer->texInfo.height));
 
 	// Vertex shader
 	rDrawState.pVertexShader = RdrShaderSystem::GetVertexShader(kScreenVertexShader);
